DE21884: Create defect in EDP does not work if the Release field is hidden.
Stop setting release and iteration to 'u' via URL parameters when their value is unscheduled.
There's nothing special about 'u' except that it is invalid for a release or iteration ref.
Leaving them off and letting them default works just as well and won't try to set the value to
'u' if one of the fields is hidden.